Arbitrary generation of well-typed expressions.
As the expression language gets larger, it is necessary to only generate
well-typed expressions in tests.
Otherwise, we end up generating far too many bad expressions and
discarding them, which makes the testing too slow.
It's still important to run some tests with the whole range of
expressions, but some, especially those that require multiple
well-typed expressions are infeasible.