Remove "Take" stream transformer.
I just added the "Take" transformer because it seemed simple, but it
turns out that it actually complicates computing the
lookbehind/bubblegum quite a bit.
Given that none of the examples I've seen benefit from it, and it could
be implemented as a fold anyway, I'm removing it.

I think the contrast between "take" and "latestN" is interesting though,
since they both seem to do the same thing but have very different
space requirements and so on.