Avalanche.FromCore: add "element prefix" and "accumulator prefix".
In Core, there are a few different namespaces: scalars, streams and
reduces.
While scalars and reduces are actually in the same namespace, the
lifetime and order of bindings changes drastically.
Avalanche merges all these into the one namespace, so we need to prefix
the stream elements and accumulators with something so they do not clash
with any existing scalar bindings.