Substitution: substitution into Let was giving up too easily.
In the case
```subst a := e in (let n := p in q)```
if neither p nor q mention a, we do not need to perform substitution
into them.
However I was only looking at q and ignoring p.

The lesson here is probably that I was being too clever by attempting to
make substitution perform slightly less work, which complicated the
implementation and led to the bug.

(It might seem weird to go to the effort to check whether "a" is mentioned
in q first, instead of just performing the substitution regardless.
However, if "n" clashes with "a" or "e", we would need to rename "n".
If "n" clashes but "q" does not mention "a", we needn't rename.)