Avalanche: convert entire Avalanche program to Statements.
We needed to add statements to pre and postcomputations anyway, so that
we could execute loops (eg array folds) outside the main loop.

However, replacing the precomps with statements alone did not make much
sense since any let bindings in there wouldn't be visible etc.

So instead, rip out the whole precomps, postcomps and replace the whole
Avalanche with a single set of statements.
I still need to add extra invariant checks, such as there can only be
one fact loop, but this should simplify the optimisation transforms we
can perform on Avalanche as a whole, since we're only dealing with a
single statement set now.

It feels like we lose a lot of structure by mingling everything together
in Statement like this, but I don't think this is a big deal. By this
stage we've already taken advantage of the structure, by performing
fusion/tupling, condensing and so on.