Alpha equivalence: better tests.
We have "free variables" implemented now, so we can write some better
properties for alpha equivalence.

If an expression is closed - it has no free variables - we can rename
any variable and the result is still alpha equivalent.

If the expression is not closed, we can only rename bound variables, not
frees.