Icicle: Source types: add property for Progress, tweaks/hacks
Add property "if it type checks, it runs". However, this required
several tweaks to the typesystem: things like "newest", "oldest" and
"fold1" only return something if there is actually data there.
Obviously we don't want to pollute every value with an implicit "null",
but we also don't want to put the burden of unboxing Maybes onto the
user.
I think this is a new modality which can be nested in the existing
universes: "possibly".