Icicle: add annotations and source positions.
Queries and expressions from the parser should be annotated with their
source positions for better error reporting.
Also, we want the type checker to be able to return a modified query
with each node annotated with its type.
So the annotation should be polymorphic.

Dum dum - I should have done this to start with. I don't know why I
didn't.