Icicle: types: introduce "possibility" for things that might fail.
For example, "newest" won't work if there are actually no entries.
Division won't work if the operand is zero.

However, dealing with these explicitly can be tedious. So instead the
"possibly modality" lets you apply operations to "potentially failing
computations", with the result being a potentially failing computation
itself.

This way we don't have to manually plumb through the failure cases, but
still get the benefits of a "real type system".