Core: make Windowing a stream transformer.
Window was a "Source" instead of a "Transformer" before.
This meant that windows could only occur at the start of a pipeline.
While this is completely valid and expressive, it makes translation from
Source to Core more complicated than necessary.
Removing the restriction does not complicate the Core translation to
Avalanche.