Core: replace "FoldPair/unpair#" primitive with saner fst and snd
I originally had pair projections implemented as a single fold/destructor:
```
unpair# :: (a -> b -> c) -> (a, b) -> c
```
which is a bit confusing, but seemed sensible for consistency: given
that we have no primitive "case" construct, everything is expressed as
folds.
Using this sort of fold function makes sense for other datatypes,
but it ends up being an unnecessary hassle for pairs and so on which
have only one constructor.