Icicle: conversion: subtle name shadowing bug, fold with possiblies
1:
Given something like
```
feature salary
~> let value  = 0
~> let value2 = value + 1
~> sum value2
```
Here, "value" does not refer to the actual input value, but instead
refers to the precomputation.
However, conversion was treating the reference to value as the input
value, which is invalid in a precomputation.

2:
"let fold" was looking at the return type to decide whether to fold as a
possibly, but should instead be looking at the fold workers.
For example
```
let fold animal = 1 : 1 ~> animal / 2
```
Here, the fold itself is definite, but the return of the overall query
is possible because of the /2.