Convert: groups with possible/option/error subqueries. break stuff
This error stuff is way too complicated. I must be doing something wrong.
Maybe the option type for division or newest/oldest is not the right
way to do it.  Adding bottom/null to Core feels dirty though, so maybe
exceptions are the way to go.

Another option is adding effects/modalities (Possibly/Definitely) to
Core: that way Array.map etc could be polymorphic in the effect they
use.

I will perservere for now, but maybe we want something slightly
higher-level than Core, either as an intermediate step, or to replace
Core.