Avalanche: separate loops over historic and new facts
When executing the feature, we also keep track of what data will be required to compute the next snapshot.

This consists of the list of individual facts that contribute
to "windowed" and "latest" features,
as well as the last values of any resumable features like reductions.

It is important that since the resumable features have already seen the
historical data, they cannot see it again.
This is why we have two separate loops, so the first loop over historical data
does not compute the resumable features:

1. Initialise variables for latest and windowed features
2. Loop through historical data, computing latest and windowed features
3. Read last values of resumable variables
4. Loop through new data, computing all features
5. Store last values of resumable variables
6. Return