Avalanche: type checking: split up into context & actual checking.
In conversion we need to reconstruct the types of things.
While it would be easier in some cases to annotate all Avalanche
statements with types, this is a rather large change that affects
everything.
Instead, it seems simpler to have conversion to Java just build up a
typechecking context of its own and reconstruct the types that way.