Rewriting reify to use dumb, simple recursion
I was using transformX etc which do the recursion, but that turned out
to be quite a lot harder because it's top down, and actually recurses
over the returned result.

It turns out you need to do all these weird gymnastics so that the
worker function doesn't modify the things it just produced.

In this case it's a lot simpler just to write the recursion out
explicitly.