Source uses VDoubles instead of VInts.
This is necessary because of polymorphism.
A literal number, LitInt i, actually has type ``forall a. Num a => a``.
This means that before type inference, we don't know whether a given
literal should be VInt or VDouble.
Instead of making every primitive that deals with numbers have to
convert between the two, we instead make it so that the evaluator uses
doubles exclusively as the number representation type.