Case with pure scrutinee and aggregate return is broken #597
Fix in two ways. One. make temporality of pattern bindings equal to result
temporality.
So this is outlawed, because the bindings 'b' and 'c' are Aggregates,
despite the scrutinee being pure.

> feature salary ~> case (0,0) | (b,c) -> (fold a = b : c ~> a) end

This makes the type checker more restrictive, but this shouldn't affect
any existing programs because they would currently fail to typecheck in
Core.

Two. Converting an Aggregate case to Core, some of the references will show up
in the precomputations. But the pattern isn't bound until post.
So take the set of names bound by the pattern, and run through the
precomputations, deferring anything that mentions the pattern names
until post. That way it will be nested beneath the pattern bindings.